EDIT CUSTOMER

+++ database/migrations/2022_05_05_113042_add_price_column_to_customer_item_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::table('customer_item', function (Blueprint $table) {
             $table->double('item_price')->nullable();
-            $table->double('quantity')->nullable();
+            $table->integer('quantity')->nullable();
             $table->tinyInteger('status')->default(0);
         });
     }
